extract event handlers from redispubsub. rename to camelCase
