+++ bigbluebutton-html5/imports/startup/server/collectionManagers/cursor.js
