+++ bigbluebutton-html5/imports/startup/server/collectionManagers/poll.js
