+++ bigbluebutton-html5/imports/startup/server/collectionManagers/shapes.js
