+++ bigbluebutton-html5/imports/startup/server/collectionManagers/slides.js
