+++ bigbluebutton-html5/imports/startup/server/index.js
@@ -1,21 +1,16 @@
 import { logger } from '/imports/startup/server/logger';
-
-import { RedisPubSub } from '/imports/startup/server/redispubsub2';
-import { EventQueue } from '/imports/startup/server/EventQueue';
-
 import '/server/server';
+import { RedisPubSub } from '/imports/startup/server/RedisPubSub';
+import { EventQueue } from '/imports/startup/server/EventQueue';
+import { clearCollections } from '/imports/startup/server/helpers';
+
+Meteor.startup(function(){
+  clearCollections();
+  logger.info('server start');
+});
 
 export const myQueue = new EventQueue();
-
-logger.info('server start');
 
 export const eventEmitter = new (Npm.require('events').EventEmitter);
 
 export const redisPubSub = new RedisPubSub();
-
-// TODO
-/*
-
--clear collections somewhere
-
- */
