+++ bigbluebutton-html5/imports/startup/server/meteorMethods/deletePrivateChatMessages.js
