+++ bigbluebutton-html5/imports/startup/server/meteorMethods/kickUser.js
@@ -1,6 +1,5 @@
-import { publish } from '../redispubsub';
-import { isAllowedTo } from '../user_permissions';
-import { appendMessageHeader } from '/imports/startup/server/helpers';
+import { isAllowedTo } from '/imports/startup/server/userPermissions';
+import { appendMessageHeader, publish } from '/imports/startup/server/helpers';
 import { redisConfig } from '/config';
 
 Meteor.methods({
