+++ bigbluebutton-html5/imports/startup/server/meteorMethods/publishVoteMessage.js
@@ -1,5 +1,5 @@
-import { publish } from '/imports/startup/server/redispubsub';
-import { isAllowedTo } from '/imports/startup/server/user_permissions';
+import { publish } from '/imports/startup/server/helpers';
+import { isAllowedTo } from '/imports/startup/server/userPermissions';
 import { appendMessageHeader } from '/imports/startup/server/helpers';
 import { Polls } from '/collections/collections';
 import { logger } from '/imports/startup/server/logger';
