+++ bigbluebutton-html5/imports/startup/server/meteorMethods/unmuteUser.js
@@ -1,22 +1,22 @@
-import { publish } from '/imports/startup/server/redispubsub';
-import { isAllowedTo } from '/imports/startup/server/user_permissions';
+import { publish } from '/imports/startup/server/helpers';
+import { isAllowedTo } from '/imports/startup/server/userPermissions';
 import { appendMessageHeader } from '/imports/startup/server/helpers';
-import { updateVoiceUser } from '/imports/startup/server/collection_methods/users';
+import { updateVoiceUser } from '/imports/startup/server/collectionManagers/users';
 import { logger } from '/imports/startup/server/logger';
 import { redisConfig } from '/config';
 
 Meteor.methods({
   // meetingId: the meetingId of the meeting the user[s] is in
-  // toMuteUserId: the userId of the user to be muted
+  // toMuteUserId: the userId of the user to be unmuted
   // requesterUserId: the userId of the requester
   // requesterToken: the authToken of the requester
-  muteUser(meetingId, toMuteUserId, requesterUserId, requesterToken) {
+  unmuteUser(meetingId, toMuteUserId, requesterUserId, requesterToken) {
     let action, message;
     action = function () {
       if (toMuteUserId === requesterUserId) {
-        return 'muteSelf';
+        return 'unmuteSelf';
       } else {
-        return 'muteOther';
+        return 'unmuteOther';
       }
     };
 
@@ -25,17 +25,17 @@ Meteor.methods({
         payload: {
           user_id: toMuteUserId,
           meeting_id: meetingId,
-          mute: true,
+          mute: false,
           requester_id: requesterUserId,
         }
       };
       message = appendMessageHeader('mute_user_request_message', message);
-      logger.info(`publishing a user mute request for ${toMuteUserId}`);
+      logger.info(`publishing a user unmute request for ${toMuteUserId}`);
       publish(redisConfig.channels.toBBBApps.users, message);
       updateVoiceUser(meetingId, {
         web_userid: toMuteUserId,
         talking: false,
-        muted: true,
+        muted: false,
       });
     }
   }
