+++ bigbluebutton-html5/imports/startup/server/userPermissions.js
