+++ bigbluebutton-html5/server/server.js
@@ -1,34 +1,29 @@
 import '/imports/startup/server/index';
-import '/imports/startup/server/collection_methods/chat';
-import '/imports/startup/server/collection_methods/cursor';
-import '/imports/startup/server/collection_methods/meetings';
-import '/imports/startup/server/collection_methods/poll';
-import '/imports/startup/server/collection_methods/presentations';
-import '/imports/startup/server/collection_methods/shapes';
-import '/imports/startup/server/collection_methods/slides';
-import '/imports/startup/server/collection_methods/users';
-import '/imports/startup/server/methods/deletePrivateChatMessages';
-import '/imports/startup/server/methods/kickUser';
-import '/imports/startup/server/methods/listenOnlyRequestToggle';
-import '/imports/startup/server/methods/muteUser';
-import '/imports/startup/server/methods/publishSwitchToNextSlideMessage';
-import '/imports/startup/server/methods/publishSwitchToPreviousSlideMessage';
-import '/imports/startup/server/methods/publishVoteMessage';
-import '/imports/startup/server/methods/sendChatMessagetoServer';
-import '/imports/startup/server/methods/setUserPresenter';
-import '/imports/startup/server/methods/unmuteUser';
-import '/imports/startup/server/methods/userLogout';
-import '/imports/startup/server/methods/userSetEmoji';
-import '/imports/startup/server/methods/validateAuthToken';
+import '/imports/startup/server/collectionManagers/chat';
+import '/imports/startup/server/collectionManagers/cursor';
+import '/imports/startup/server/collectionManagers/meetings';
+import '/imports/startup/server/collectionManagers/poll';
+import '/imports/startup/server/collectionManagers/presentations';
+import '/imports/startup/server/collectionManagers/shapes';
+import '/imports/startup/server/collectionManagers/slides';
+import '/imports/startup/server/collectionManagers/users';
+import '/imports/startup/server/meteorMethods/deletePrivateChatMessages';
+import '/imports/startup/server/meteorMethods/kickUser';
+import '/imports/startup/server/meteorMethods/listenOnlyRequestToggle';
+import '/imports/startup/server/meteorMethods/muteUser';
+import '/imports/startup/server/meteorMethods/publishSwitchToNextSlideMessage';
+import '/imports/startup/server/meteorMethods/publishSwitchToPreviousSlideMessage';
+import '/imports/startup/server/meteorMethods/publishVoteMessage';
+import '/imports/startup/server/meteorMethods/sendChatMessagetoServer';
+import '/imports/startup/server/meteorMethods/setUserPresenter';
+import '/imports/startup/server/meteorMethods/unmuteUser';
+import '/imports/startup/server/meteorMethods/userLogout';
+import '/imports/startup/server/meteorMethods/userSetEmoji';
+import '/imports/startup/server/meteorMethods/validateAuthToken';
 import '/imports/startup/server/EventQueue';
 import '/imports/startup/server/helpers';
 import '/imports/startup/server/logger';
 import '/imports/startup/server/publish';
-import '/imports/startup/server/redispubsub';
-import '/imports/startup/server/redispubsub2';
-import '/imports/startup/server/user_permissions';
-
-
-
-// import { redisPubSub } from '/imports/startup/server/index';
-// console.log("in server I check if I can publish", redisPubSub.constructor());
+import '/imports/startup/server/eventHandlers';
+import '/imports/startup/server/RedisPubSub';
+import '/imports/startup/server/userPermissions';
